take out logic of adding events
